[simi-toolkit-core] refactor: Move the simi-initializer source code to the "simi-toolkit-core" module
