+++ simi-core/simi-toolkit-core/pom.xml
@@ -17,4 +17,10 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
+    <dependencies>
+        <dependency>
+            <groupId>com.simi</groupId>
+            <artifactId>simi-common-utils</artifactId>
+        </dependency>
+    </dependencies>
 </project>
