+++ simi-core/simi-toolkit-core/src/main/java/com/simi/toolkit/core/SimiControllerUtils.java
@@ -1,7 +1,7 @@
-package com.simi.common.util.file;
+package com.simi.toolkit.core;
 
 
-import com.simi.common.util.file.support.controller.ControllerPattern;
+import com.simi.toolkit.core.constants.ControllerPattern;
 import lombok.experimental.UtilityClass;
 
 import java.util.Optional;
