+++ simi-core/simi-toolkit-core/src/main/java/com/simi/toolkit/core/constants/ControllerPattern.java
@@ -1,4 +1,4 @@
-package com.simi.common.util.file.support.controller;
+package com.simi.toolkit.core.constants;
 
 import java.util.regex.Pattern;
