+++ simi-core/simi-toolkit-core/src/main/java/com/simi/toolkit/core/domain/yaml/ProjectInfo.java
@@ -1,4 +1,4 @@
-package com.simi.common.util.file.support.yaml;
+package com.simi.toolkit.core.domain.yaml;
 
 import lombok.Data;
