+++ simi-core/simi-toolkit-core/src/main/java/com/simi/toolkit/core/domain/yaml/WriteInfo.java
@@ -1,5 +1,6 @@
-package com.simi.common.util.file.support.yaml;
+package com.simi.toolkit.core.domain.yaml;
 
+import com.simi.toolkit.core.domain.enums.BackupEnum;
 import lombok.Data;
 
 import java.util.List;
@@ -9,7 +10,7 @@ public class WriteInfo {
     private String write;
     private String read;
     private String type;
-    private String backup=BackupEnum.CURRENT.getValue();
+    private String backup= BackupEnum.CURRENT.getValue();
     private Boolean once;
     private List<String> rpRuleList;
     private StringList activeEnvList;
