+++ simi-core/simi-toolkit-core/src/main/java/com/simi/toolkit/core/exception/InitException.java
@@ -1,4 +1,4 @@
-package com.simi.common.util.file.support;
+package com.simi.toolkit.core.exception;
 
 import lombok.NoArgsConstructor;
